// src/app/models/ticket.ts
export interface Ticket {
    id?: any;                        
    openingDate?: string;    
    closingDate?: string;    
    priority: string;       
    status: string;         
    title: string;          
    observations: string;   
    technical: any;      
    customer: any;       
    technicianName: string; 
    customerName: string;   
}